Move location unlock one level up to MainScreen
This enables location unlock while on other tabs (not just MapScreen)
