+++ src/components/Map.tsx
@@ -1,27 +1,15 @@
 import React, { useEffect, useRef, useState } from "react";
 import { Platform, View } from "react-native";
 import MapView, { MAP_TYPES, Marker, UrlTile } from "react-native-maps";
-import { useRecoilState, useRecoilValue } from "recoil";
-import { locationSetup } from "../location/location";
-import {
-  currentEventState,
-  currentLocation,
-  activeQuestsState,
-} from "../recoil/atom";
+import { useRecoilValue } from "recoil";
+import { currentLocation, activeQuestsState } from "../recoil/atom";
 import { useQuery } from "react-query";
 import { FontAwesome5, AntDesign } from "@expo/vector-icons";
 import { TILE_URL_TEMPLATE } from "@env";
 import MarkerCard from "./MarkerCard";
-import { QuestItem, QuestParticipation, QuestStatus } from "../client";
+import { QuestItem } from "../client";
 import { IconButton } from "react-native-paper";
-import { questsWithinUnlockRadius } from "../location/locationUnlock";
-import { currentUser } from "../recoil/atom";
-import {
-  getQuestItems,
-  getUserEventQuests,
-  updateQuestParticipation,
-} from "../api/quests";
-import { sendNotification } from "../notifications/notifications";
+import { getQuestItems } from "../api/quests";
 
 enum MarkerType {
   POI = "POI",
@@ -32,35 +20,14 @@ const Map = () => {
   const MIN_ZOOM_LEVEL = 17;
   const MAX_ZOOM_LEVEL = 21;
   const map = useRef<MapView | null>(null);
-  const [location, setLocation] = useRecoilState(currentLocation);
+  const location = useRecoilValue(currentLocation);
   const [focusUserLocation, setFocusUserLocation] = useState(true);
   const [selectedMarker, setSelectedMarker] = useState<null | {
     id: string;
     markerType: MarkerType;
   }>(null);
   const [animateToCoordinate, setAnimateToCoordinate] = useState(false);
-  const user = useRecoilValue(currentUser);
-  const [activeQuests, setActiveQuests] = useRecoilState(activeQuestsState);
-  const currentEvent = useRecoilValue(currentEventState);
-
-  // Simple "lock" to prevent multiple location unlocks for the same target
-  // (which could cause things like multiple notifications)
-  const [performingLocationUnlock, setPerformingLocationUnlock] =
-    useState(false);
-
-  const { data: unstartedQuests, refetch: refetchUnstartedQuests } = useQuery<
-    QuestParticipation[],
-    Error
-  >(["unstartedQuests", user, currentEvent], () => {
-    if (user && currentEvent) {
-      return getUserEventQuests(
-        user.id,
-        currentEvent.id,
-        QuestStatus.UNSTARTED
-      );
-    }
-    return unstartedQuests;
-  });
+  const activeQuests = useRecoilValue(activeQuestsState);
 
   const { data: questItems } = useQuery<QuestItem[], Error>(
     ["questItems", activeQuests],
@@ -73,54 +40,6 @@ const Map = () => {
       ).flat();
     }
   );
-
-  useEffect(() => {
-    // Caution! State variables will be outdated when accessed from the location update callback,
-    // since they where frozen when the callback subscription was created (in `locationSetup`).
-    // This is why we only update the state, and use a seperate useEffect to react to the change
-    locationSetup(setLocation);
-  }, []);
-
-  useEffect(() => {
-    if (unstartedQuests && user && currentEvent && !performingLocationUnlock) {
-      setPerformingLocationUnlock(true);
-      const unlockedQuests = questsWithinUnlockRadius(
-        location.coords,
-        unstartedQuests.map((qp) => qp.quest)
-      );
-      if (unlockedQuests.length == 0) {
-        setPerformingLocationUnlock(false);
-        return;
-      }
-      Promise.all(
-        unlockedQuests.map((quest) => {
-          sendNotification(
-            "Quest Unlocked ✨",
-            `You unlocked '${quest.title}'`
-          );
-          return updateQuestParticipation(
-            user.id,
-            quest.id,
-            QuestStatus.ACTIVE
-          );
-        })
-      )
-        .then(() => {
-          return getUserEventQuests(
-            user.id,
-            currentEvent.id,
-            QuestStatus.ACTIVE
-          );
-        })
-        .then((activeQuests) => {
-          setActiveQuests(activeQuests);
-          return refetchUnstartedQuests();
-        })
-        .then(() => {
-          setPerformingLocationUnlock(false);
-        });
-    }
-  }, [location, unstartedQuests]);
 
   useEffect(() => {
     if (focusUserLocation && location) {
