+++ src/recoil/atom.ts
@@ -1,4 +1,4 @@
-import { atom } from "recoil";
+import { atom, TransactionInterface_UNSTABLE } from "recoil";
 import { LocationObject } from "expo-location";
 import { User, Event, QuestParticipation } from "../client";
 
@@ -19,6 +19,18 @@ export const currentLocation = atom({
   key: "currentLocation", // unique ID (with respect to other atoms/selectors)
   default: defaultLocation, // default value (aka initial value)
 });
+
+export const resetUserStateTransaction: (
+  i: TransactionInterface_UNSTABLE
+) => (...args: readonly unknown[]) => void =
+  ({ reset }) =>
+  () => {
+    reset(currentUser);
+    reset(currentEventState);
+    reset(activeQuestsState);
+    reset(userQuestsProgressState);
+    reset(questScreenVisibleQuestState);
+  };
 
 export const currentUser = atom<User | null>({
   key: "currentUser",
