+++ src/screens/QuestLogScreen.tsx
@@ -5,23 +5,20 @@ import { QuestStackParamList } from "./QuestNavigator";
 import { Pressable } from "react-native";
 import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
 import {
-  currentEventState,
   currentUser,
   questScreenVisibleQuestState,
   userQuestsProgressState,
   activeQuestsState,
 } from "../recoil/atom";
-import { QuestParticipation, QuestStatus } from "../client";
-import { useQuery } from "react-query";
-import { getUserEventQuests, getUserQuestProgress } from "../api/quests";
+import { QuestParticipation } from "../client";
+import { getUserQuestProgress } from "../api/quests";
 
 const QuestLogScreen = ({
   navigation,
 }: NativeStackScreenProps<QuestStackParamList, "Quest log">) => {
   const user = useRecoilValue(currentUser);
 
-  const currentEvent = useRecoilValue(currentEventState);
-  const setActiveQuests = useSetRecoilState(activeQuestsState);
+  const activeQuests = useRecoilValue(activeQuestsState);
 
   const [userQuestsProgress, setUserQuestsProgress] = useRecoilState(
     userQuestsProgressState
@@ -31,35 +28,23 @@ const QuestLogScreen = ({
     questScreenVisibleQuestState
   );
 
-  const { data: userQuests } = useQuery<QuestParticipation[], Error>(
-    ["userQuests", currentEvent],
-    () =>
-      user && currentEvent
-        ? getUserEventQuests(user.id, currentEvent.id, QuestStatus.ACTIVE)
-        : []
-  );
-
-  // TODO: Would want to use onSuccess in useQuery instead, but it doesn't seem to be reactive enough
   useEffect(() => {
-    if (userQuests) {
-      setActiveQuests(userQuests);
-      if (user) {
-        userQuests.forEach((q) => {
-          getUserQuestProgress(user.id, q.quest.id).then((p) => {
-            setUserQuestsProgress((existing) => ({
-              ...existing,
-              [q.quest.id]: p,
-            }));
-          });
+    if (activeQuests && user) {
+      activeQuests.forEach((q) => {
+        getUserQuestProgress(user.id, q.quest.id).then((p) => {
+          setUserQuestsProgress((existing) => ({
+            ...existing,
+            [q.quest.id]: p,
+          }));
         });
-      }
+      });
     }
-  }, [userQuests]);
+  }, [activeQuests]);
 
   const DATA: { title: string; data: QuestParticipation[] }[] = [
     {
       title: "Active Quests",
-      data: userQuests ?? [],
+      data: activeQuests ?? [],
     },
   ];
